layer.conf: avoid dangling openvswitch bbappend
Issue: LIN1022-2957

There are no hard dependencies on meta-virtualization in this layer
but there is a bbappend for openvswitch, should meta-virtualization be
present. This means that builds with this layer but without
meta-virtualization will have a dangling bbappends ERROR/WARNING:

    ERROR: No recipes in default available for:
    .../layers/intel-x86/recipes-networking/openvswitch/openvswitch_git.bbappend

Here we make use of BBFILES_DYNAMIC to setup allow the bbappend to be
applied *only* if the meta-virtualization layer is present. We also
set meta-virtualization as a LAYERRECOMMENDS to make the relationship
between the two layers explicit.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>
Signed-off-by: Xiangyu Chen <xiangyu.chen@windriver.com>